run backup: fix container live CLI
After the fix to allow the run-backup to be executed in a non-interactive
terminal, the feature to drop into the live container that runs the backup
was broken as it was missing the `-it` docker flag. Now the docker command
is refactored intor 2 components - the run command with the run flags and
the docker flags that concern the container. This allows to insert the
`-it` docker flag when the desired runtime is a live container.